refactor(core): 重构节点心跳及配置相关代码
- 将阈值设置改为直接传递，简化 AlertServiceImpl 中的阈值处理逻辑
- 新增 NodeConfig 配置类，拆分节点配置，支持节点身份及心跳频率设置
- JobNodeHeartbeat 使用 NodeConfig 替代 Leader 配置，调整节点状态字段名称为 nodeStatus
- JobNodeVO 新增节点角色 nodeRole 字段，完善节点信息描述
- 调整 DefaultTaskDispatcher 中节点 ID 和实例 ID 获取，统一使用 JobLockGuard.INSTANCE_ID 常量
- 修正 DefaultTaskDispatcher 中连续失败次数查询，使用 Optional 防止空指针
- FlowInstanceLifecycleService 修改保存逻辑，保存实体转换后的 DTO
- 修正 WorkflowUserVO 和 JobDagVO，增加多租户 tenantId 字段支持
- 调整 WebhookOutboxSubscriber 中事件类型处理，避免空指针异常
- MetricsOutboxSubscriber 新增作业执行相关事件类型常量定义
- ScanTask 接口增加默认扫描间隔常量
- 代码格式优化，删除无用导入，调整日志打印变量名称一致性

+++ ydsz-cronjob/ydsz-cronjob-server/src/main/java/com/njydsz/cronjob/server/core/handler/GlueJobHandler.java
@@ -78,7 +78,7 @@ import com.njydsz.cronjob.server.service.schedule.GlueCodeService;
 @ConditionalOnMissingBean(GlueJobHandler.class)
 public class GlueJobHandler implements JobHandler {
   /** 脚本执行超时（秒） */
-  private static final long SCRIPT_TIMEOUT_SECONDS = 300;
+  private static final int SCRIPT_TIMEOUT_SECONDS = 300;
 
 
   /** Bean 名称，dispatcher 在 jobType=GLUE 时路由到此 handler */
